+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Fixed `Thumbnail` focus state style.
 -   Fixed `Thumbnail` fallback placement.
 -   Fixed `Thumbnail` fill height style.
+-   Fixed `TabProvider` and `ProgressTrackerProvider` state on unmout/remount.
 
 ## [1.0.2][] - 2021-01-18